+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 13.2.0
+
+- Expose `attributeGroupType` for `AttributesFilterItemWithValues` on `/filters` endpoint
+- Support `includeShopSizeRuns` parameter on `/filters` endpoint
+
 ### 13.1.0
 
 - Now supports CloudVault's token based authentication.